Add `gitkeep [branch name]` command
Add a command to delete all local branches except the one with the same
name as the argument.

This command allows us to type, for instance,

```
gitkeep develop2
```

To delete all of our local branches, except the one named "develop2".

Caveat: If a different branch is currently checked out when running this
command, the deletion of the currently-checkout-out-branch will fail.

---

After a while, local branches sometimes build up when they're not
regularly deleted, causing the Git UI we use to track development (GitX)
to have a long list of local branch names on the left side of the
screen.

It isn't a terrible thing, but it's nice to keep things tidy and it's
nice to be able to do it quickly, when you need to, rather than
individually deleting all the outdated local branches.